test-app/requirements.txt
aiida-core~=1.1.1
panel~=0.8.0
tornado==5.1.1